Fixed type casting

+++ ProductCatalog.Service/Entities/Product.cs
@@ -10,12 +10,11 @@ namespace ProductCatalog.Service.Entities
     {
         public int Id { get; set; }
         public required string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int StockQuantity { get; set; }
         public bool IsActive { get; set; }
         public DateTime CreatedAt { get; set; }
         public int CategoryId { get; set; }
-
         public ProductCategory Category { get; set; } = null!;
 
     }
